【优化】1、设置默认收发地址0xffff 2、优化保存参数

+++ driver/eeprom.h
@@ -3,7 +3,6 @@
 #include "ebox_core.h"
 
 #define EEPROM_STARTADDR 0x1000  
-uint8_t EEPROM_Read(uint16_t Addr,uint8_t *RxBuffer,uint8_t Length);
-uint8_t EEPROM_Write(uint16_t Addr,uint8_t *RxBuffer,uint8_t Lenth);
-
+uint8_t EEPROM_Read(uint16_t Addr,uint8_t *RxBuffer,int Length);
+uint8_t EEPROM_Write(uint16_t Addr,uint8_t *RxBuffer,int Lenth);
 #endif
